+++ ricaun.Revit.Github/GithubRequestService.cs
@@ -1,5 +1,9 @@
 ﻿using ricaun.Revit.Github.Services;
 using System;
+using System.Linq;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
 
 namespace ricaun.Revit.Github
 {
@@ -21,10 +25,12 @@ namespace ricaun.Revit.Github
         /// </summary>
         /// <param name="user"></param>
         /// <param name="repo"></param>
+        [MethodImpl(MethodImplOptions.NoInlining)]
         public GithubRequestService(string user, string repo)
         {
+            var assembly = Assembly.GetCallingAssembly();
             this.githubBundleService = new GithubBundleService(user, repo);
-            this.pathBundleService = new PathBundleService();
+            this.pathBundleService = new PathBundleService(assembly);
             this.downloadBundleService = new DownloadBundleService();
         }
 
@@ -46,9 +52,33 @@ namespace ricaun.Revit.Github
             var bundleModel = githubBundleService.GetBundleModelLatest();
 
             if (bundleModel is not null)
-                downloadBundleService.DownloadBundleAsync(folder, bundleModel.DownloadUrl);
+                downloadBundleService.DownloadBundle(folder, bundleModel.DownloadUrl);
 
             return bundleModel;
+        }
+
+        /// <summary>
+        /// Initialize
+        /// </summary>
+        /// <returns></returns>
+        public Task Initialize()
+        {
+            var task = Task.Run(async () =>
+                {
+                    var bundleModels = await githubBundleService.GetBundleModelsAsync();
+                    if (bundleModels.Any() == false) return;
+                    var bundleModel = bundleModels.FirstOrDefault();
+                    Console.WriteLine(bundleModel);
+
+                    if (pathBundleService.TryGetPath(out string folder))
+                    {
+                        await downloadBundleService.DownloadBundleAsync(folder, bundleModel.DownloadUrl);
+                    }
+
+                    Console.WriteLine($"Initialize Finish");
+                });
+
+            return task;
         }
 
     }