+++ CHANGELOG.md
@@ -5,17 +5,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] / 2022-05-21
-### ToDo
-- Async Methods - Problem
 ### Features
 - [x] Download Bundle From Github
 - [ ] Improve Library Files
+- [ ] Select Download Version
 #### GithubRequestService
 - [x] Clear
 - [x] Remove Github Data 
 - [x] Add PathBundle
 - [x] Add XML comment
+### Fixed
+- Fix Async Methods - Problem
 ### Added
+- Add Async in all Services
 - Add BundleModel (Version, DownloadUrl, Body, DataTime)
 - Add DownloadBundleService
 - Add GithubBundleService